+++ src/main/webapp/web.xml
@@ -13,7 +13,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>allUsers</servlet-name>
-        <servlet-class>internet.shop.controllers.UsersController</servlet-class>
+        <servlet-class>internet.shop.controllers.user.UsersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allUsers</servlet-name>
@@ -21,7 +21,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>registration</servlet-name>
-        <servlet-class>internet.shop.controllers.RegistrationController</servlet-class>
+        <servlet-class>internet.shop.controllers.user.RegistrationController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>registration</servlet-name>
@@ -29,7 +29,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>deleteUser</servlet-name>
-        <servlet-class>internet.shop.controllers.DeleteUserController</servlet-class>
+        <servlet-class>internet.shop.controllers.user.DeleteUserController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteUser</servlet-name>
@@ -37,15 +37,23 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>allProducts</servlet-name>
-        <servlet-class>internet.shop.controllers.ProductsController</servlet-class>
+        <servlet-class>internet.shop.controllers.product.ProductsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allProducts</servlet-name>
         <url-pattern>/products/all</url-pattern>
     </servlet-mapping>
     <servlet>
+        <servlet-name>adminAllProducts</servlet-name>
+        <servlet-class>internet.shop.controllers.product.AdminProductsController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>adminAllProducts</servlet-name>
+        <url-pattern>/products/adminAll</url-pattern>
+    </servlet-mapping>
+    <servlet>
         <servlet-name>createProduct</servlet-name>
-        <servlet-class>internet.shop.controllers.CreateProductController</servlet-class>
+        <servlet-class>internet.shop.controllers.product.CreateProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>createProduct</servlet-name>
@@ -53,7 +61,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>deleteProduct</servlet-name>
-        <servlet-class>internet.shop.controllers.DeleteProductController</servlet-class>
+        <servlet-class>internet.shop.controllers.product.DeleteProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteProduct</servlet-name>
@@ -61,7 +69,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>shoppingCart</servlet-name>
-        <servlet-class>internet.shop.controllers.ShoppingCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.ShoppingCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>shoppingCart</servlet-name>
@@ -69,7 +77,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>addProductToShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.AddProductToShopCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.AddProductToShopCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>addProductToShopCart</servlet-name>
@@ -77,11 +85,43 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>removeProductFromShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.RemoveProductFromShopCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.RemoveProductFromShopCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>removeProductFromShopCart</servlet-name>
         <url-pattern>/removeProductFromShopCart</url-pattern>
+    </servlet-mapping>
+    <servlet>
+        <servlet-name>CreateOrder</servlet-name>
+        <servlet-class>internet.shop.controllers.order.CreateOrderController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>CreateOrder</servlet-name>
+        <url-pattern>/CreateOrder</url-pattern>
+    </servlet-mapping>
+    <servlet>
+        <servlet-name>DeleteOrder</servlet-name>
+        <servlet-class>internet.shop.controllers.order.DeleteOrderController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>DeleteOrder</servlet-name>
+        <url-pattern>/DeleteOrder</url-pattern>
+    </servlet-mapping>
+    <servlet>
+        <servlet-name>OrderInfo</servlet-name>
+        <servlet-class>internet.shop.controllers.order.OrderInfoController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>OrderInfo</servlet-name>
+        <url-pattern>/OrderInfo</url-pattern>
+    </servlet-mapping>
+    <servlet>
+        <servlet-name>AllOrders</servlet-name>
+        <servlet-class>internet.shop.controllers.order.OrderController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>AllOrders</servlet-name>
+        <url-pattern>/orders/all</url-pattern>
     </servlet-mapping>
     <servlet>
         <servlet-name>InjectData</servlet-name>
